extracting some tests
helper function tests should now be closer to the code
